+++ src/sll/api/string.c
@@ -183,8 +183,8 @@ static void _object_to_string(sll_object_t* a,sll_string_t* o){
 					return;
 				}
 				const sll_object_type_data_t* dt=*(sll_current_runtime_data->tt->dt+a->t-SLL_MAX_OBJECT_TYPE-1);
-				if (dt->fn.str){
-					sll_object_t* v=sll_execute_function(dt->fn.str,&a,1,0);
+				if (dt->fn[SLL_OBJECT_FUNC_STRING]){
+					sll_object_t* v=sll_execute_function(dt->fn[SLL_OBJECT_FUNC_STRING],&a,1,0);
 					sll_object_t* str=sll_operator_cast(v,sll_static_int[SLL_OBJECT_TYPE_STRING]);
 					GC_RELEASE(v);
 					sll_string_increase(o,str->dt.s.l);