+++ src/sll/include/sll/_object_func.h
@@ -1,0 +1,73 @@
+#ifndef __SLL__OBJECT_FUNC_H__
+#define __SLL__OBJECT_FUNC_H__ 1
+/**
+ * \flags subgroup
+ * \name Functions
+ * \group object
+ * \subgroup object-func
+ * \desc Docs!
+ */
+
+
+
+/**
+ * \flags macro var
+ * \name SLL_OBJECT_FUNC_INIT
+ * \group object
+ * \subgroup object-func
+ * \desc Docs!
+ * \type sll_object_function_index_t
+ */
+#define SLL_OBJECT_FUNC_INIT 0
+
+
+
+/**
+ * \flags macro var
+ * \name SLL_OBJECT_FUNC_DELETE
+ * \group object
+ * \subgroup object-func
+ * \desc Docs!
+ * \type sll_object_function_index_t
+ */
+#define SLL_OBJECT_FUNC_DELETE 1
+
+
+
+/**
+ * \flags macro var
+ * \name SLL_OBJECT_FUNC_COPY
+ * \group object
+ * \subgroup object-func
+ * \desc Docs!
+ * \type sll_object_function_index_t
+ */
+#define SLL_OBJECT_FUNC_COPY 2
+
+
+
+/**
+ * \flags macro var
+ * \name SLL_OBJECT_FUNC_STRING
+ * \group object
+ * \subgroup object-func
+ * \desc Docs!
+ * \type sll_object_function_index_t
+ */
+#define SLL_OBJECT_FUNC_STRING 3
+
+
+
+/**
+ * \flags macro var
+ * \name SLL_MAX_OBJECT_FUNC
+ * \group object
+ * \subgroup object-func
+ * \desc Docs!
+ * \type sll_object_function_index_t
+ */
+#define SLL_MAX_OBJECT_FUNC SLL_OBJECT_FUNC_STRING
+
+
+
+#endif